Fixed imports after moved 'functions' folder into common, and moved requestStatuses into one index.js file
