+++ src/common/components/DetailsTile/index.js
@@ -15,9 +15,9 @@ import { selectMovieDetails } from
     "../../../features/MovieDetailsPage/slices/movieDetailsSlice";
 import { StyledProfileIcon } from "../StyledProfileIcon";
 import { StyledVideoIcon } from "../StyledVideoIcon";
-import { getYear } from "../../../functions/getYear";
-import { getImageUrl } from "../../../functions/getImageUrl.js"
-import { checkIsValidImageUrl } from "../../../functions/checkIsValidImageUrl.js";
+import { getYear } from "../../../common/functions/getYear";
+import { getImageUrl } from "../../../common/functions/getImageUrl.js"
+import { checkIsValidImageUrl } from "../../../common/functions/checkIsValidImageUrl.js";
 import { INVALID_IMAGE_URL } from "../../constants/config/INVALID_IMAGE_URL.js";
 import { IMAGE_WIDTH } from "../../constants/config/IMAGE_WIDTH.js";
