+++ src/features/MovieDetailsPage/components/MainContent/index.js
@@ -10,7 +10,7 @@ import { PageContent } from "../../../../common/components/PageContent";
 import { BannerContent, MovieTitle, BannerMainInfo, Banner } from "./styled";
 import { toPerson } from "../../../../routes";
 import { useNavigationToPage } from "../../../../useNavigation";
-import { getImageUrl } from "../../../../functions/getImageUrl";
+import { getImageUrl } from "../../../../common/functions/getImageUrl";
 import { BACKDROP_WIDTH } from "../../../../common/constants/config/BACKDROP__WIDTH";
 
 export const MainContent = () => {
