+++ src/features/MovieDetailsPage/components/index.js
@@ -5,9 +5,9 @@ import { MainContent } from "./MainContent";
 import { fetchMoviesCreditsList, selectMovieCreddits, selectMoviesCreditsListFetchStatus } from "../slices/moviesCredditsListSlice";
 import { Loader } from "../../../common/components/Loader";
 import { Error } from "../../../common/components/Error";
-import { loadingStatus } from "../../../common/constants/requestStatuses/loadingStatus";
-import { errorStatus } from "../../../common/constants/requestStatuses/errorStatus";
-import { successStatus } from "../../../common/constants/requestStatuses/successStatus";
+import { loadingStatus } from "../../../common/constants/requestStatuses";
+import { errorStatus } from "../../../common/constants/requestStatuses";
+import { successStatus } from "../../../common/constants/requestStatuses";
 import { fetchMoviesGenres, selectMoviesGenresFetchStatus } from "../../../common/slices/moviesGenresSlice";
 import { useParams } from "react-router-dom";
 import { NoResults } from "../../../common/components/NoResultsPage";
