+++ src/features/MoviesListPage/components/index.js
@@ -4,8 +4,8 @@ import { useLocation } from 'react-router-dom';
 import { selectPopularMoviesFetchStatust, fetchPopularMovies, selectCurrentPage, selectTotalPages } from '../slices/popularMoviesSlice';
 import { searchMovies } from '../../../common/slices/searchMoviesSlice';
 import { fetchMoviesGenres, selectMoviesGenresFetchStatus } from '../../../common/slices/moviesGenresSlice';
-import { loadingStatus } from '../../../common/constants/requestStatuses/loadingStatus';
-import { errorStatus } from '../../../common/constants/requestStatuses/errorStatus';
+import { loadingStatus } from '../../../common/constants/requestStatuses';
+import { errorStatus } from '../../../common/constants/requestStatuses';
 import { Loader } from '../../../common/components/Loader';
 import { Error } from '../../../common/components/Error';
 import { MainContent } from './MainContent';
