+++ src/features/MoviesListPage/slices/popularMoviesSlice.js
@@ -1,7 +1,7 @@
 import { createSlice } from "@reduxjs/toolkit";
-import { successStatus } from "../../../common/constants/requestStatuses/successStatus";
-import { loadingStatus } from "../../../common/constants/requestStatuses/loadingStatus";
-import { errorStatus } from "../../../common/constants/requestStatuses/errorStatus";
+import { successStatus } from "../../../common/constants/requestStatuses";
+import { loadingStatus } from "../../../common/constants/requestStatuses";
+import { errorStatus } from "../../../common/constants/requestStatuses";
 
 export const popularMoviesSlice = createSlice({
     name: "popularMovies",
